Refactor MTCLog file structure and handlers
